// ui/src/queries/types/TodoItem.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: TodoItem
// ====================================================

export interface TodoItem {
  __typename: "Todo";
  id: string;
  text: string;
  isChecked: boolean;
}